Refactor: refactor `app` folder in "Model-AI" according to the recoomended structure of Nest.JS.
