+++ src/app/app.controller.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { AppController } from '@controllers/app.controller';
-import { AppService } from '../services/app.service';
+import { AppController } from '@app/app.controller';
+import { AppService } from './app.service';
 
 describe('AppController', () => {
   let appController: AppController;
