+++ src/app/app.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { OpenAIService } from '../openai/services/openai.service';
+import { OpenAIService } from '../openai/openai.service';
 
 @Injectable()
 export class AppService {
