+++ src/main.ts
@@ -7,9 +7,10 @@ async function bootstrap() {
   const app = await NestFactory.create(AppModule);
 
   app.enableCors({
-    origin:'*',
+    origin:'http://localhost:3000', 
+    credentials:true
   });
-  
+
   
   await app.listen(3000);
 }
